fix: Use the AniBridge-defined StrEnum for sync_rules templates
